+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVV8P6D1NFM6NMV0GTQG5JQ7"
+        "01JVVFS9S8GHVJJ32RVZC85KD2"
     )
     r"""The id of the resource"""