+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVV8P6D1EB42TE29YQ1205A8"
+        "01JVVFS9S7HD3XZH3N1HF2Z2Y5"
     )
     r"""The id of the resource"""