+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-22T05:50:26.701Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-22T07:54:28.337Z")
     r"""The date and time the resource was last updated"""